Move testUtils into a sibling dir
