+++ test/resources/InvoiceItems.spec.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var stripe = require('../testUtils').getSpyableStripe();
+var stripe = require('../../testUtils').getSpyableStripe();
 var expect = require('chai').expect;
 
 describe('InvoiceItems Resource', function() {
