+++ testUtils/.eslintrc.js
@@ -1,0 +1,14 @@
+module.exports = {
+    "env": {
+        "mocha": true
+    },
+    "plugins": [
+        "chai-friendly"
+    ],
+    "rules": {
+        "no-loop-func": "off",
+        "no-sync": "off",
+        "no-unused-expressions": 0,
+        "chai-friendly/no-unused-expressions": 2
+    }
+};
